+++ packages/core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Refactor resource paths and update documentation indexing. (`89d0d78f`)
 - Add new images and plugin metadata. (`37969b6e`)
 
 ## 0.1.3 - 2026-07-25